Refactor: extract some functions to other files and clean code
