+++ src/components/ItemList/index.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
-import { updateItemDate } from '../Utils/firestore';
-import { isWithin24hours, latestInterval } from 'components/Utils/helpers';
-import calculateEstimate from 'lib/estimates';
+import { isWithin24hours, isDateValid } from 'components/Utils/helpers';
+import { markProductPurchased } from 'components/Utils/firestore';
+
 function ItemList({
   itemName,
   docId,
@@ -12,7 +12,6 @@ function ItemList({
 }) {
   const [isChecked, setIsChecked] = useState(false);
   const [isDisabled, setIsDisabled] = useState(false);
-  const [token] = useState(() => window.localStorage.getItem('tcl18-token'));
   const [formattedDate, setFormattedDate] = useState(0);
 
   useEffect(() => {
@@ -29,37 +28,6 @@ function ItemList({
       setIsDisabled(true);
     }
   }, [formattedDate, lastPurchasedDate]);
-
-  const isDateValid = (date) => {
-    if (date) return true;
-  };
-  const markProductPurchased = (
-    id,
-    lastPurchasedDateMillis,
-    nextPurchaseEstimatedByUser,
-    actualNumberOfPurchases,
-    lastEstimateNextPurchase,
-  ) => {
-    const daysLatestInterval = latestInterval(
-      lastPurchasedDateMillis,
-      actualNumberOfPurchases,
-      nextPurchaseEstimatedByUser,
-    );
-
-    const estimatedNextPurchase = calculateEstimate(
-      lastEstimateNextPurchase,
-      daysLatestInterval,
-      actualNumberOfPurchases,
-    );
-
-    updateItemDate(token, id, actualNumberOfPurchases, estimatedNextPurchase)
-      .then(() => {
-        console.log('product updated');
-      })
-      .catch((error) => {
-        console.error(error);
-      });
-  };
 
   const handleCheckbox = (event) => {
     setIsChecked(!isChecked);
